# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Internal SQL API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "integer"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "title": "TokenResponse"
      },
      "def-1": {
        "type": "object",
        "properties": {
          "q": {
            "type": "string",
            "enum": [
              "get_prefix",
              "get_estado_responsable",
              "update_estado"
            ]
          },
          "params": {
            "type": "object",
            "additionalProperties": true
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 100
          },
          "timeoutMs": {
            "type": "integer",
            "minimum": 100,
            "maximum": 60000,
            "default": 8000
          }
        },
        "required": [
          "q"
        ],
        "title": "ReadQueryBody"
      },
      "def-2": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "ok",
          "data"
        ],
        "title": "ReadQueryResponse"
      },
      "def-3": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "enum": [
              "create_order"
            ]
          },
          "args": {
            "type": "object",
            "additionalProperties": true
          },
          "timeoutMs": {
            "type": "integer",
            "minimum": 100,
            "maximum": 60000,
            "default": 8000
          }
        },
        "required": [
          "op"
        ],
        "title": "CmdRunBody"
      },
      "def-4": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "ok"
        ],
        "title": "CmdRunResponse"
      }
    }
  },
  "paths": {
    "/auth/token": {
      "post": {
        "summary": "Emite un JWT RS256 (solo desarrollo)",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientId": {
                    "type": "string"
                  },
                  "clientSecret": {
                    "type": "string"
                  },
                  "scope": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": [
                      "read.sql"
                    ]
                  }
                },
                "required": [
                  "clientId",
                  "clientSecret"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/read/query": {
      "post": {
        "summary": "Ejecuta una consulta predefinida (lectura)",
        "tags": [
          "Read"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-1"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          }
        }
      }
    },
    "/cmd/run": {
      "post": {
        "summary": "Ejecuta una operación predefinida (comando)",
        "tags": [
          "Command"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-3"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-4"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}